+++ copier_template_tester/_config.py
@@ -21,8 +21,9 @@ def _validate_config(config: dict) -> None:  # type: ignore[type-arg]
 def load_config(base_dir: Path) -> dict:  # type: ignore[type-arg]
     """Read the ctt config from `CWD`."""
     cfg_path = base_dir / 'ctt.toml'
-    if cfg_path.is_file():
-        config: dict = tomllib.loads(cfg_path.read_text())  # type: ignore[type-arg]
-        _validate_config(config)
-        return config
-    raise ValueError(f'No configuration file found. Expected: {cfg_path.absolute()}')  # pragma: no cover # noqa: EM102
+    if not cfg_path.is_file():  # pragma: no cover
+        msg = f'No configuration file found. Expected: {cfg_path.absolute()}'
+        raise ValueError(msg)
+    config: dict = tomllib.loads(cfg_path.read_text())  # type: ignore[type-arg]
+    _validate_config(config)
+    return config