+++ tests/helpers.py
@@ -5,6 +5,7 @@ from pathlib import Path
 from typing import Generator
 
 from beartype import beartype
+from corallium.file_helpers import read_lines
 from corallium.log import get_logger
 from pytestshellutils.shell import Subprocess
 from pytestshellutils.utils.processes import ProcessResult
@@ -71,3 +72,15 @@ def temporary_git_dir(shell: Subprocess, *, source_dir: Path | None = None) -> G
             add_commit(shell, cwd=working_dir)
 
         yield working_dir
+
+
+@beartype
+def reset_path_in_test_answers() -> None:
+    """Replace the absolute path written by copier.."""
+    answers_path = TEST_DATA_DIR / 'copier_demo/.ctt/no_all/.copier-answers.testing_no_all.yml'
+    lines = (
+        '_src_path: ../../copier_demo' if line.startswith('_src_path') else line
+        for line in read_lines(answers_path)
+        if line
+    )
+    answers_path.write_text('\n'.join(lines) + '\n')